Merge pull request #33436 from MrDoomBringer/DropItLikeItsHottest
For too many years, Cargonia has been oppressed, depressed, and re-pressed, by the slow and cumbersome movement of the supply shuttle. After hours of begging and pleading around on coderbus, I can present to you the ultimate cargo utility.

This update adds room for a certain level of progression with cargo, allowing for extra money to be spent on extra-fast delivery times. The drop pod itself picks a random, unblocked turf in the cargobay, then crashes down upon it, creating a small explosion and damaging some tiles.

The explosion exists to act as a secondary act of balance: the drop pod is intended to be used for a few VIP deliveries. Ordering in bulk will still require use of the shuttle, unless you are prepared to fix the shit out of an explosion-riddled cargo.

Because of the nature of the mechanic, it is very easy to spend a lot of money very quickly. For this reason, a separate console was created: so that it could be placed in the safety of the Quartermaster's office, away from the sausage fingers of incompetent greyshits and rookie cargotechs.