Refactors and nerfs Fleshmend (#34458)
🆑 Xhuis
balance: Fleshmend can no longer be stacked several times at once.
balance: Fleshmend no longer heals changelings who are on fire, and heals burn damage half as quickly as brute and oxygen damage.
code: Fleshmend has been refactored into a status effect.
/🆑

Fleshmend is one of the abilities that makes changelings unkillable combat juggernauts with essentially no counterplay, so I decided to change it:

    First of all, refactored it into a status effect that cannot stack, meaning that only one instance of Fleshmend can be active at once
    Instead of healing 10/10/10 brute/burn/oxy damage, it now heals 10/5/10, emphasizing changelings' weakness to fire
    If a changeling is on fire, Fleshmend will not function at all

This keeps Fleshmend powerful for healing in and out of combat, but adds counterplay to it and gives changelings something to fear: fire. With an extinguisher, they still need to manage their inventory to use it, and even with a fire-immune suit and helmet using the fire on them as a weapon will be unable to use their Fleshmend trump card to outheal any damage they take.

I'll be keeping an eye on changelings after this is merged and adjusting accordingly.