[Ready] Re-adds M-90gl to Nuke Ops uplink (#39152)
I thought this weapon was interesting enough to re-add as a Nuke Ops weapon. In the course of making this PR I went back to some of the PRs from when it was removed from the uplink to get a good sense of what was seen as wrong with it at the time. I have done my best to deal with these issues and balance it properly.
In terms of power, the weapon basically fits between the C-20r and the L6 SAW. It's more portable than a SAW, but generally less powerful, and costs just as much to purchase and outfit. The weapon fires half as fast as a SAW and the bullets do less damage, just barely critting an unarmored person with a full burst. Targets with even the relatively weak Security Officer body armor can withstand a burst from it, and people wearing Bulletproof Armor can take multiple bursts before going down. The grenade launcher is strong, but slow to reload, and expensive to use a lot.

For people who remember the old one, this is basically the same thing but significantly more expensive. I also changed the grenade shells to be purchasable individually for 2 TC each instead of buying a box of 4 for 5 TC. There's also no ridiculously cheap bundle this time around, that can stay gone.