Adds a warning and single retry to byond membership lookup (#86765)
## About The Pull Request

A few people are having issues getting byond membership features
disabled even though they're a byond member. This is *likely* due to
byond server troubles, and according to lummox the lookup proc should
return null when a connection issue happens. So I've put some handling
in there for that case as well as a single retry.

:cl:
fix: Byond membership lookup should now warn you when it fails due to a
connection failure.
/:cl: