[i ded] Jean shorts are correct clothes for the death sandwich (#87280)
## About The Pull Request

If you know the underlying reference behind the death sandwich, you
would think the jean shorts from the clothesmate would allow you to eat
it right. This is not the case, because it checks for exactly one
specific type. This is inconsistent with the source material, as there
is more than one kind of clothing in this game that includes jean
shorts.

## Why It's Good For The Game

i ded why reference not work like it do in source

## Changelog

:cl:
qol: All forms of jean shorts now count for eating the death sandwich
right
/:cl: