Make roundstart book publishing on bird possible. (#87843)
## About The Pull Request
Adds a book binder to the birdshot library in the spot the book scanner
used to be and moves the scanner right next to the library computer on
the desk.

## Why It's Good For The Game
Curator shouldn't have to wait for research to be able to put books into
the archive.
## Changelog
:cl: Goat
map: The library's scanner on Birdshot is now close enough to connect to
the computer and was also given a book binder.
/:cl: