Reduces required crayons for crayon bounty (#87848)
## About The Pull Request

This reduces the bounty size of the crayon civilian pack.

Instead of 24 crayons, it only needs 8. Same payout, less quantity.
## Why It's Good For The Game

Crayons are not easily produced or scavenged. Even if a map contains 2
full crayon packs in the Library, it still isn't enough to fulfill a
single bounty for crayons.

This new deal will get you a full payout with a full pack plus one
_mystery crayon._ Go rob the clown or detective or something if you need
the last one. Get that money!