Fixes mechs with crowbar-like tools holding firelocks open at a distance (#87907)
## About The Pull Request
Changes firelock code to register signals on the atom holding the
crowbar, not necessarily the mob that initiated the action. Since mobs
inside mechs don't technically move, signals listening for the mob's
movement wouldn't fire until the mob exits the mech. Registering the
signal on the crowbar owner (mech in this case) solves the issue.

This makes no change to mobs holding crowbars, as the mob is considered
the crowbar owner in that case, and is thus handled identically as
before.
## Why It's Good For The Game

Prevents this:


https://github.com/user-attachments/assets/77150360-c1c2-46a0-8943-3b80fbd5a9e8
## Changelog
:cl:
fix: Firelocks opened by a mech will correctly close when the mech moves
out of range.
/:cl: