crate chaos crate position randomizes again (#88011)
## About The Pull Request
moves the crate chaos randomizer marker to the far corner of the domain.
this fixes the crate always spawning in one place, somehow. running
theory is that the map didn't initialize enough before the crate
randomizer was placed, so it never found the encrypted crate so it could
never randomize its position.

moving it to the corner makes it so it only loads after most of the map
does, so it can shuffle the crate around. i tested it

![image](https://github.com/user-attachments/assets/82861120-9375-4800-8c58-008916391903)

## Why It's Good For The Game
woe, bitrunners must now interact with the map gimmick again

## Changelog

:cl:
fix: Crate Chaos's encrypted cache is now randomly placed again.
/:cl:

Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>