Meow and Purr emotes are no longer usable by all carbons (#88313)
## About The Pull Request

Currently all carbons are able to meow and purr using the new Felinid
tongue exclusive emotes. I'm changing how it works so it properly denies
anyone without Felinid tongues from using the emotes. Mimes also will
not be able to use the emote sounds so they stay silent. Lastly, there
was a delay for the emotes on their first use where the player had to
use them twice for it to work and I've fixed that, and I also removed
some redundant text warnings for when you can't use the emote.

Edit: The Purr emote has been tweaked so that it isn't a visible emote
and deaf players can't see it when it's used. Mimes can't purr anymore
because it has no visible components so they shouldn't be able to.

## Why It's Good For The Game

Every carbon meowing and purring is really bad, and it was only approved
to work for Felinids. This PR will rectify that. Multiple fixes and
tweaks to how this emote should function are also huge positives.

:cl:
fix: Only players with Felinid tongues get to meow and purr for real
this time, and other carbons cannot
fix: Mimes can no longer meow out loud and have their own mime message
fix: Stopped delay on meow and purr emotes where they had to be used
twice to work the first time they're used
del: Removed redundant text warning for players that can't use the meow
or purr emotes
tweak: Purring is no longer audible to deaf players
tweak: Players with vow of silence can't purr anymore
/:cl: