Updates the Icebox Phonebooth alarm to actually be inside the booth (#88335)
## About The Pull Request

![image](https://github.com/user-attachments/assets/101bd911-c7ae-465a-99a6-f7a91b6c01e5)

A very light remapping, a few cosmetic improvements such as snowy
plating and some space by the door so it isn't flush against rockwalls.
Most notably, the air alarm is actually on the inside now, so it reads
the room inside instead of telling you that the Icemoon is indeed
ice-cold. Also has all-access because would-be-callers might want to
shut it up.
## Why It's Good For The Game
The alarm being on the outside and reading the outside hurt me. Now it's
a tiny bit nicer and cleaner.

Except, you know. Still a frozen hellscape out there.
## Changelog
:cl:
fix: fixed the Icebox Phonebooth air alarm being on the outside, thus
triggering because the planet is, indeed, cold. It is now inside and
all-access so that callers can turn it off when they decide the phone's
more important than their health and safety.
/:cl: